streamlit_gender_classification/requirements.txt
streamlit==0.88.0
joblib==1.0.1
Pillow==8.3.2
scikit-image==0.18.3
numpy==1.19.5
tensorflow==2.6.0
tensorflow-estimator==2.6.0
tensorflow-hub==0.12.0
opencv-python-headless
opencv-python==4.5.3.56
keras==2.6.0
Keras-Preprocessing==1.1.2
scikit-image==0.18.3
scipy==1.7.1
h5py==3.1.0
